Move search components back to dialog
