+++ source/dialog/components/SearchBar.js
