+++ source/dialog/components/SearchLayout.js
