+++ source/dialog/components/SearchResults.js
