+++ source/dialog/containers/SearchResults.js
